Bug Fixes
- WMI is no more used to check Windows architecture (32 or 64-bits).
Script is faster to start now.

SetValue()
- FIX: If Key do not exist, Value was not created. Now the script checks
for the existence of the Key and create it (if necessary).
- FIX: Function did not return result value. We cant validate if the key
was written or not.
- Apprimored way to detect Typ_ if not informed one. (Now it can detect
binary values).
- Now REG_BINARY supports write values directly in hex.
Example usage will be included soon.
- Now the function returns the exactly error code if it is out from
possible ones. (Line 434)

Delete()
- Now the function returns the exactly error code if it is out from
possible ones. (Line 529)

TestPath()
- FIX: If user try to delete a Key top level from HKEY_USERS, function
will exit. (They're equivalents to HKCR and HKCU)

ConvertType()
- Now supports "REG_SZ", "REG_EXPAND_SZ", etc.
"S", "X", etc. was mantained for compatibility mode with original script
from Joey.

DeleteKey()
- FIX: Key was not being excluded. Ssubkeyname was receiving a null
value (s2_ was not defined).